backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.4.0
dj-database-url